pyproject.toml
[project]
name = "finance-purple-agent"
version = "0.1.0"
description = "Baseline Purple Agent for Finance Green Agent Benchmark"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "pydantic>=2.12.5",
    "uvicorn>=0.38.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]
